Several fixes
1) Teleporting to dead player shouldn't throw the activator out of the level
2) Only one player at a time is able to revive the other
3) Player who is giving the treatment is totally frozen during the operation
4) All revive-related text is readable now
5) Aug charge in "Team" page of "Stats" is shown properly now
6) Shield charge replaces health only when it's higher than 0 in multiplayer HUD

+++ DoomRPG/scripts/Menu.c
@@ -1028,7 +1028,7 @@ void DrawStatsMenu()
         // Aug Battery
         PrintSprite("AUGBATT", 0, 286.1, 228.1, 0.05);
         SetFont("BIGFONT");
-        HudMessage("%d%%", PlayerPtr->Augs.Battery);
+        HudMessage("%d%%", (int)PlayerPtr->Augs.Battery);
         EndHudMessage(HUDMSG_PLAIN, 0, "Yellow", 306.1, 228.0, 0.05);
 
         // Toxicity
